+++ app/src/main/java/moji/deliverytracker/MoneyCalculator.kt
@@ -4,20 +4,20 @@ import java.math.BigDecimal
 import java.math.RoundingMode
 
 object MoneyCalculator {
-    fun commissionAmount(total: Int, commissionPercent: Float): Int {
+    fun commissionAmount(total: Long, commissionPercent: Float): Long {
         if (total <= 0 || commissionPercent <= 0f) return 0
         val totalBD = BigDecimal(total)
         val percentBD = BigDecimal(commissionPercent.toString())
         return totalBD.multiply(percentBD)
             .divide(BigDecimal(100), 0, RoundingMode.HALF_UP)
-            .toInt()
+            .toLong()
     }
 
-    fun netIncome(total: Int, commissionPercent: Float): Int {
+    fun netIncome(total: Long, commissionPercent: Float): Long {
         return total - commissionAmount(total, commissionPercent)
     }
 
-    fun balance(netIncome: Int, totalPaid: Int): Int {
+    fun balance(netIncome: Long, totalPaid: Long): Long {
         return netIncome - totalPaid
     }
 }
